fix block size computation when size greater than max uint
Reviewed-on: https://gerrit.mozy.lab.emc.com/86956

+++ mordor/tar.cpp
@@ -212,7 +212,7 @@ void parsePaxAttributes(const std::string& paxAttrs, Mordor::TarEntry& entry)
 
 inline long long blockSize(long long size)
 {
-    return (size + BLOCK_SIZE - 1) & ~(BLOCK_SIZE - 1);
+    return (size + BLOCK_SIZE - 1) & ~((long long)BLOCK_SIZE - 1);
 }
 
 inline long long padSize(long long size)
